Itamar branch

+++ back/controller/vehicle_page.js
@@ -1,4 +1,5 @@
 const vehicleInfo = require("../models/index").VehicleInfo
+const { uploadPhoto, deletePhoto, updatePhoto } = require('../controller/photo_controller');
 
 const getVehicleInfo = async(req,res)=>{
 	const vehicleID = req.query.uuid
@@ -19,8 +20,13 @@ const getVehicleInfo = async(req,res)=>{
 const createVehicleInfo = async(req, res) => {
   const {  VehicleType, Depth, Width , Height } = req.body;
 	const MoverID = req.userId;
+  let PhotoUrl = '';
+  
   try {
-      const result = await vehicleInfo.create({MoverID, VehicleType, Depth, Width, Height});
+      if (req.file) {
+        PhotoUrl = await uploadPhoto(req.file);
+      }
+      const result = await vehicleInfo.create({MoverID, VehicleType, Depth, Width, Height, PhotoUrl});
       if (result) {
         res.status(201).json({ msg: 'VehicleInfo created successfully', vehicleInfo: result});
       } else {
@@ -31,11 +37,14 @@ const createVehicleInfo = async(req, res) => {
     }
 }
 const updateVehicleInfo = async (req, res) => {
-  const { MoverID, VehicleType, Depth, Width, Height } = req.body;
+  const { MoverID, VehicleType, Depth, Width, Height, PhotoUrl } = req.body;
   const vehicleID = req.query.uuid; // Assuming vehicleID is passed in the URL
   try {
+      if (req.file) {
+          PhotoUrl = updatePhoto(req.file, PhotoUrl);
+      }
       const [affectedRows, updatedRows] = await vehicleInfo.update(
-          { MoverID, VehicleType, Depth, Width, Height },
+          { MoverID, VehicleType, Depth, Width, Height, PhotoUrl},
           { where: { uuid: vehicleID }, returning: true}
       );
 
@@ -52,12 +61,22 @@ const updateVehicleInfo = async (req, res) => {
 const deleteVehicleInfo = async (req, res) => {
   const vehicleID = req.query.uuid;
   try {
-    const result = await vehicleInfo.destroy({ where: { uuid: vehicleID } });
-    if (result) {
-      res.status(200).json({ message: 'VehicleInfo deleted successfully' });
-    } else {
-      res.status(404).json({ message: 'VehicleInfo not found' });
+    const vehicleInformation = await vehicleInfo.findByPk(vehicleID);
+    if (!vehicleInformation) {
+      return res.status(404).json({ message: 'VehicleInfo not found' });
     }
+    else{
+      if(vehicleInformation.PhotoUrl){
+        deletePhoto(vehicle.PhotoUrl);
+      }
+      const result = await vehicleInfo.destroy({ where: { uuid: vehicleID } });
+      if (result) {
+        res.status(200).json({ message: 'VehicleInfo deleted successfully' });
+      } else {
+        res.status(404).json({ message: 'VehicleInfo not found' });
+      }
+    }
+    
   } catch (error) {
     res.status(500).json({ message: 'Internal Server Error', error: error.message });
   }
@@ -67,16 +86,48 @@ const getVehiclesByMoverId = async (req, res) => {
 	const moverID = req.userId;
   try {
     const vehicles = await vehicleInfo.findAll({ where: { MoverID: moverID } });
+    if (!vehicles) {
+      return res.status(404).json({ message: 'No Vehicles found' });
+    }
     return res.status(200).json({ message: "success", vehicles });
   } catch (error) {
     return res.status(500).json({ msg: "Internal Error", error: error.message });
   }
 };
+
+const deleteVehiclePhoto = async (req, res) => {
+  const vehicleID = req.query.uuid;
+  try {
+    const vehicle = await User.findOne({ where: { vehicleID } });
+    if (!vehicle) {
+      return res.status(404).json({ message: 'User not found' });
+    }
+    else if (vehicle.PhotoUrl) {
+      await deletePhoto(user.PhotoUrl);
+    }
+    else {
+      return res.status(409).json({ message: 'No Vehicle Photo found' });
+    }
+    const [affectedRows, updatedRows] = await User.update(
+      { PhotoUrl: '' }, { where: { uuid }, returning: true });
+    if (affectedRows > 0) {
+      res.status(200).json({ message: 'Vehicle Photo deleted successfully', user: updatedRows[0] });
+    }
+    else {
+      res.status(500).json({ message: 'Failed to delete Vehicle Photo' });
+    }
+  }
+  catch (error) {
+    res.status(500).json({ message: "Internal Server Error", error: error.message });
+  }
+};
+
 module.exports = {
     getVehicleInfo,
     createVehicleInfo,
     updateVehicleInfo,
     deleteVehicleInfo,
-    getVehiclesByMoverId
-};
+    getVehiclesByMoverId,
+    deleteVehiclePhoto
+  };
